feat(mailbox): store and return init_peer_id and timestamp

+++ src/spell/modules/spell/spell/src/kv/collection.rs
@@ -43,17 +43,20 @@ pub fn list_pop_string(key: &str) -> StringValue {
         get.bind(1, key)?;
         get.bind(2, key)?;
         let string = read_string(&mut get, 1)?;
-        let list_index = get.read::<i64>(3)?;
 
-        let mut delete = db.prepare(
-            r#"
+        if string.is_some() {
+            let list_index = get.read::<i64>(3)?;
+
+            let mut delete = db.prepare(
+                r#"
             DELETE FROM kv
                 WHERE key = ? AND list_index = ?
         "#,
-        )?;
-        delete.bind(1, key)?;
-        delete.bind(2, list_index)?;
-        delete.next()?;
+            )?;
+            delete.bind(1, key)?;
+            delete.bind(2, list_index)?;
+            delete.next()?;
+        }
 
         string
     };
